requirements.txt
gradio
requests
bs4
python_dotenv
langchain
langchain_core
langchain_google_genai
pinecone
langchain_community
unstructured[pdf]